fix(ci): clear the 90% coverage gate for Glade-only mirror builders
Split the pure regrid math into mirror/regrid.py (tested, counted);
omit the builder I/O modules that stream Glade sources CI cannot have;
cover jcm.data.remote fully via an injected huggingface_hub stand-in.
