+++ .coveragerc
@@ -1,6 +1,15 @@
 [run]
 omit =
     *_test.py
+    # Offline data-mirror builders: they stream multi-GB Glade sources
+    # (GMTED DEM, ERA5, input4MIPs) that cannot exist in CI. Their pure
+    # math (SSO statistics, regridding, registry) is covered via
+    # jcm/data/mirror/mirror_test.py; the I/O drivers run only on Glade.
+    jcm/data/mirror/sso.py
+    jcm/data/mirror/era5_land.py
+    jcm/data/mirror/ozone.py
+    jcm/data/mirror/emissions.py
+    jcm/data/mirror/bundles.py
     # Two-moment microphysics process functions are ported but not yet fully
     # wired into the column orchestrator — remove these entries once Phase 5b
     # lands the full ECHAM6 sequence (see issue #341).
