+++ jcm/data/mirror/bundles.py
@@ -35,7 +35,10 @@ from __future__ import annotations
 
 import numpy as np
 import xarray as xr
-from scipy.spatial import cKDTree
+
+from jcm.data.mirror.regrid import (conservative_to_gaussian,  # noqa: F401
+                                    fill_nearest, gaussian_latlon,
+                                    interp_to)
 
 AMIP_ROOT = ("/glade/campaign/cesm/cesmdata/input4MIPs_raw/input4MIPs/"
              "CMIP7/CMIP/PCMDI/PCMDI-AMIP-1-1-10")
@@ -62,51 +65,7 @@ CLIMO_TIME = np.array([np.datetime64(f"2014-{m:02d}-01")
                        for m in range(1, 13)])
 
 
-def gaussian_latlon(nlat: int):
-    lats = np.rad2deg(np.arcsin(np.polynomial.legendre.leggauss(nlat)[0]))
-    return lats, np.arange(2 * nlat) * 360.0 / (2 * nlat)
 
-
-def interp_to(da: xr.DataArray, lats, lons) -> xr.DataArray:
-    """Bilinear regrid with periodic longitude wrap; lat clamped at ends."""
-    latn, lonn = da.dims[-2], da.dims[-1]
-    if float(da[latn][0]) > float(da[latn][-1]):
-        da = da.isel({latn: slice(None, None, -1)})
-    dlon = float(da[lonn][1] - da[lonn][0])
-    wrapped = xr.concat(
-        [da.isel({lonn: -1}).assign_coords(
-            {lonn: float(da[lonn][0]) - dlon}),
-         da,
-         da.isel({lonn: 0}).assign_coords(
-             {lonn: float(da[lonn][-1]) + dlon})], dim=lonn)
-    # constant extension to the poles so Gaussian lats beyond the source's
-    # first/last row interpolate instead of going NaN
-    if float(wrapped[latn][0]) > -90.0:
-        wrapped = xr.concat(
-            [wrapped.isel({latn: 0}).assign_coords({latn: -90.0}), wrapped],
-            dim=latn)
-    if float(wrapped[latn][-1]) < 90.0:
-        wrapped = xr.concat(
-            [wrapped, wrapped.isel({latn: -1}).assign_coords({latn: 90.0})],
-            dim=latn)
-    out = wrapped.interp({latn: lats, lonn: lons}, method="linear")
-    return out.rename({latn: "lat", lonn: "lon"})
-
-
-def _fill_nearest(field: np.ndarray, lats, lons) -> np.ndarray:
-    """Fill NaNs (land in ocean products) with the nearest valid value."""
-    glon, glat = np.meshgrid(np.deg2rad(lons), np.deg2rad(lats))
-    xyz = np.stack([np.cos(glat) * np.cos(glon),
-                    np.cos(glat) * np.sin(glon), np.sin(glat)], -1)
-    out = field.copy()
-    for t in range(field.shape[0]):
-        bad = ~np.isfinite(field[t])
-        if not bad.any():
-            continue
-        tree = cKDTree(xyz[~bad])
-        _, idx = tree.query(xyz[bad], workers=-1)
-        out[t][bad] = field[t][~bad][idx]
-    return out
 
 
 def _monthly_clim(da: xr.DataArray, era: str) -> xr.DataArray:
@@ -121,33 +80,6 @@ def _to_lonlat(da2d: xr.DataArray) -> tuple:
                                   if d not in ("lat", "lon"))
     return dims, da2d.transpose(*dims).values
 
-
-def conservative_to_gaussian(field: np.ndarray, src_lats, src_lons,
-                             lats, lons) -> np.ndarray:
-    """Area-weighted binning of a regular-grid flux onto Gaussian cells.
-
-    Each source cell contributes its cos(lat)-weighted value to the target
-    cell containing its center — conservative in the flux-density sense,
-    which is what per-m² emissions need (bilinear would smear point
-    sources and lose mass).
-    """
-    lat_edges = np.concatenate([[-90.0], 0.5 * (lats[1:] + lats[:-1]),
-                                [90.0]])
-    nlat, nlon = lats.size, lons.size
-    dlon = 360.0 / nlon
-    lat_bin = np.clip(np.searchsorted(lat_edges, src_lats) - 1, 0, nlat - 1)
-    lon_bin = ((np.asarray(src_lons) - (lons[0] - dlon / 2)) // dlon
-               ).astype(int) % nlon
-    flat = (lat_bin[:, None] * nlon + lon_bin[None, :]).ravel()
-    w = np.cos(np.deg2rad(src_lats))[:, None].repeat(len(src_lons), 1).ravel()
-    wsum = np.bincount(flat, weights=w, minlength=nlat * nlon)
-    lead = field.shape[:-2]
-    out = np.empty(lead + (nlat, nlon))
-    for idx in np.ndindex(lead):
-        num = np.bincount(flat, weights=w * field[idx].ravel(),
-                          minlength=nlat * nlon)
-        out[idx] = (num / np.maximum(wsum, 1e-30)).reshape(nlat, nlon)
-    return out
 
 
 _EMIS_SPECIES = ("so2", "bc", "oc")
@@ -204,7 +136,7 @@ def build_forcing(era5_path: str, era: str, lats, lons,
     tos = xr.open_dataset(TOS).tos
     sic = xr.open_dataset(SICONC).siconc
     sst_c = _monthly_clim(tos, era)
-    sst = _fill_nearest(sst_c.values, sst_c.lat.values, sst_c.lon.values)
+    sst = fill_nearest(sst_c.values, sst_c.lat.values, sst_c.lon.values)
     sst_da = xr.DataArray(sst + 273.15, dims=("time", "lat", "lon"),
                           coords={"time": CLIMO_TIME,
                                   "lat": sst_c.lat, "lon": sst_c.lon})
