+++ jcm/data/mirror/mirror_test.py
@@ -8,8 +8,8 @@ import unittest
 
 import numpy as np
 
-from jcm.data.mirror.bundles import (conservative_to_gaussian,
-                                     gaussian_latlon, interp_to)
+from jcm.data.mirror.regrid import (conservative_to_gaussian,
+                                    gaussian_latlon, interp_to)
 from jcm.data.mirror.sso import finalize
 from jcm.data.mirror.registry import build_registry, write_registry
 
@@ -90,6 +90,18 @@ class RegridTest(unittest.TestCase):
         tgt_int = (out.mean(axis=1) * w_tgt).sum()
         self.assertAlmostEqual(tgt_int / src_int, 1.0, places=6)
 
+    def test_fill_nearest_takes_closest_valid_value(self):
+        from jcm.data.mirror.regrid import fill_nearest
+        lats = np.array([0.0, 1.0])
+        lons = np.array([0.0, 1.0, 2.0])
+        field = np.array([[[1.0, np.nan, 3.0],
+                           [1.0, 1.0, np.nan]]])
+        out = fill_nearest(field, lats, lons)
+        self.assertTrue(np.isfinite(out).all())
+        self.assertEqual(out[0, 1, 2], 3.0)     # nearest valid is (0, 2)
+        # untouched cells stay identical
+        self.assertEqual(out[0, 0, 0], 1.0)
+
     def test_interp_to_wraps_longitude(self):
         import xarray as xr
         src = xr.DataArray(
@@ -118,14 +130,27 @@ class ResolveDataPathTest(unittest.TestCase):
                          "/local/x.nc")
         self.assertIsNone(runners._resolve_data_path(None))
 
-    def test_fetch_is_cache_first(self):
-        # a warm cache must resolve with local_files_only (no network)
-        try:
-            import huggingface_hub  # noqa: F401
-        except ImportError:
-            self.skipTest("huggingface_hub not installed")
+    @staticmethod
+    def _fake_hub(download):
+        """sys.modules stand-in for huggingface_hub (works without it)."""
+        import sys
+        import types
         from unittest import mock
 
+        class LocalEntryNotFoundError(Exception):
+            pass
+
+        hub = types.ModuleType("huggingface_hub")
+        hub.hf_hub_download = download
+        errors = types.ModuleType("huggingface_hub.errors")
+        errors.LocalEntryNotFoundError = LocalEntryNotFoundError
+        hub.errors = errors
+        patcher = mock.patch.dict(sys.modules, {
+            "huggingface_hub": hub, "huggingface_hub.errors": errors})
+        return patcher, LocalEntryNotFoundError
+
+    def test_fetch_is_cache_first(self):
+        # a warm cache must resolve with local_files_only (no network)
         from jcm.data import remote
 
         calls = []
@@ -134,10 +159,35 @@ class ResolveDataPathTest(unittest.TestCase):
             calls.append(kw)
             return "/cache/hit"
 
-        with mock.patch("huggingface_hub.hf_hub_download", side_effect=fake):
+        patcher, _ = self._fake_hub(fake)
+        with patcher:
             self.assertEqual(remote.fetch("bundles/x.nc"), "/cache/hit")
         self.assertEqual(len(calls), 1)
         self.assertTrue(calls[0]["local_files_only"])
+
+    def test_fetch_cache_miss_goes_online_then_errors_helpfully(self):
+        from jcm.data import remote
+
+        def cold_then_ok(**kw):
+            if kw.get("local_files_only"):
+                raise err("not cached")
+            return "/downloaded"
+
+        patcher, err = self._fake_hub(cold_then_ok)
+        with patcher:
+            self.assertEqual(
+                remote.bundle_file("t63", "terrain.nc"), "/downloaded")
+
+        def always_fails(**kw):
+            if kw.get("local_files_only"):
+                raise err2("not cached")
+            raise ConnectionError("no internet")
+
+        patcher, err2 = self._fake_hub(always_fails)
+        with patcher:
+            with self.assertRaisesRegex(FileNotFoundError,
+                                        "prefetch on a login node"):
+                remote.fetch("bundles/t63/terrain.nc")
 
     def test_list_paths_resolve_elementwise(self):
         from unittest import mock
